refactor: camera manager
